requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
aiohttp>=3.8.5
requests>=2.28.0
ruff>=0.7.3
pre-commit>=4.0.0

# Database (PRP-003) - PostgreSQL only
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
greenlet>=3.0.0

# AI/ML (PRP-002, PRP-006, PRP-007)
openai>=1.0.0
pgvector>=0.2.0
sentence-transformers>=2.2.0

# Redis (PRP-002: LESSONS cache)
redis>=5.0.0
aioredis>=2.0.0

# Scheduling (PRP-008)
apscheduler>=3.10.0

# Browser testing (PRP-016 interactive location system)
playwright>=1.40.0
pytest-playwright>=0.4.3

# E2E Testing with User Client (PRP-013)
pyrogram[fast]>=2.0.0
tgcrypto>=1.2.0

# Web Search (PRP-005)
duckduckgo-search>=6.0.0

# External Tools (PRP-009)
httpx>=0.27.0
google-search-results>=2.4.2

# Image Generation (PRP-016 tile system)
replicate>=0.15.0
pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.3.0

# Discord Integration (PRP-XXX: Discord Integration)
discord.py>=2.3.0
discord-py-interactions>=5.0.0

# Analytics & Observability (PRP-012)
prometheus-client>=0.19.0
langsmith>=0.1.0
structlog>=23.0.0
aiohttp>=3.8.5
